+++ NEWS.md
@@ -1,6 +1,4 @@
 # rbiips 0.11.0
 
 * rbiips is now a standalone package, including biips c++ libraries
-
-
-
+* in pmmh: added `proposal` as argument with default = `"prior"`